requirements.txt
asgiref==3.2.5
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
Django==3.0.4
django-ratelimit==2.0.0
django-user-agents==0.4.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
idna==2.9
Pillow==7.0.0
PyJWT==1.7.1
python-dateutil==2.8.1
python-memcached==1.59
pytz==2019.3
qrcode==6.1
requests==2.23.0
six==1.14.0
sqlparse==0.3.1
ua-parser==0.10.0
urllib3==1.25.8
user-agents==2.1
